+++ src/main/java/tobyspring/splearn/adapter/integration/DummyEmailSender.java
@@ -3,7 +3,7 @@ package tobyspring.splearn.adapter.integration;
 import org.springframework.context.annotation.Fallback;
 import org.springframework.stereotype.Component;
 import tobyspring.splearn.application.required.EmailSender;
-import tobyspring.splearn.domain.Email;
+import tobyspring.splearn.domain.shared.Email;
 
 @Component
 @Fallback
