+++ src/main/java/tobyspring/splearn/application/provided/MemberFinder.java
@@ -1,6 +1,6 @@
 package tobyspring.splearn.application.provided;
 
-import tobyspring.splearn.domain.Member;
+import tobyspring.splearn.domain.member.Member;
 
 /**
  * 회원을 조회한다
