+++ src/main/java/tobyspring/splearn/application/provided/MemberRegister.java
@@ -1,8 +1,8 @@
 package tobyspring.splearn.application.provided;
 
 import jakarta.validation.Valid;
-import tobyspring.splearn.domain.Member;
-import tobyspring.splearn.domain.MemberRegisterRequest;
+import tobyspring.splearn.domain.member.Member;
+import tobyspring.splearn.domain.member.MemberRegisterRequest;
 
 /**
  * 회원 등록과 관련된 기능을 제공한다
