+++ src/main/java/tobyspring/splearn/application/required/MemberRepository.java
@@ -2,8 +2,8 @@ package tobyspring.splearn.application.required;
 
 import java.util.Optional;
 import org.springframework.data.repository.Repository;
-import tobyspring.splearn.domain.Email;
-import tobyspring.splearn.domain.Member;
+import tobyspring.splearn.domain.shared.Email;
+import tobyspring.splearn.domain.member.Member;
 
 /**
  * 회원 정보를 저장하거나 조회한다
