+++ src/main/java/tobyspring/splearn/domain/member/MemberDetail.java
@@ -1,4 +1,4 @@
-package tobyspring.splearn.domain;
+package tobyspring.splearn.domain.member;
 
 import jakarta.persistence.Entity;
 import java.time.LocalDateTime;
@@ -6,6 +6,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
+import tobyspring.splearn.domain.AbstractEntity;
 
 @Entity
 @Getter
