+++ src/main/java/tobyspring/splearn/domain/member/MemberRegisterRequest.java
@@ -1,4 +1,4 @@
-package tobyspring.splearn.domain;
+package tobyspring.splearn.domain.member;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Size;
