+++ src/main/java/tobyspring/splearn/domain/member/MemberStatus.java
@@ -1,4 +1,4 @@
-package tobyspring.splearn.domain;
+package tobyspring.splearn.domain.member;
 
 public enum MemberStatus {
     PENDING, ACTIVE, DEACTIVATED
