+++ src/main/java/tobyspring/splearn/domain/member/PasswordEncoder.java
@@ -1,4 +1,4 @@
-package tobyspring.splearn.domain;
+package tobyspring.splearn.domain.member;
 
 public interface PasswordEncoder {
     String encode(String password);
