+++ src/main/resources/META-INF/orm.xml
@@ -15,7 +15,7 @@
         </attributes>
     </mapped-superclass>
 
-    <entity class="tobyspring.splearn.domain.Member">
+    <entity class="tobyspring.splearn.domain.member.Member">
         <table name="member">
             <unique-constraint name="UK_MEMBER_EMAIL_ADDRESS">
                 <column-name>email_address</column-name>
@@ -36,7 +36,7 @@
         </attributes>
     </entity>
 
-    <embeddable class="tobyspring.splearn.domain.Email">
+    <embeddable class="tobyspring.splearn.domain.shared.Email">
         <attributes>
             <basic name="address">
                 <column name="email_address" nullable="false" length="150"/>
