+++ src/test/java/tobyspring/splearn/SplearnTestConfiguration.java
@@ -4,7 +4,7 @@ import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import tobyspring.splearn.application.required.EmailSender;
 import tobyspring.splearn.domain.MemberFixture;
-import tobyspring.splearn.domain.PasswordEncoder;
+import tobyspring.splearn.domain.member.PasswordEncoder;
 
 @TestConfiguration
 public class SplearnTestConfiguration {
