+++ src/test/java/tobyspring/splearn/application/provided/MemberFinderTest.java
@@ -9,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Transactional;
 import tobyspring.splearn.SplearnTestConfiguration;
-import tobyspring.splearn.domain.Member;
+import tobyspring.splearn.domain.member.Member;
 import tobyspring.splearn.domain.MemberFixture;
 
 @SpringBootTest
