+++ src/test/java/tobyspring/splearn/application/provided/MemberRegisterManualTest.java
@@ -17,10 +17,10 @@ import tobyspring.splearn.application.MemberModifyService;
 import tobyspring.splearn.application.MemberQueryService;
 import tobyspring.splearn.application.required.EmailSender;
 import tobyspring.splearn.application.required.MemberRepository;
-import tobyspring.splearn.domain.Email;
-import tobyspring.splearn.domain.Member;
+import tobyspring.splearn.domain.shared.Email;
+import tobyspring.splearn.domain.member.Member;
 import tobyspring.splearn.domain.MemberFixture;
-import tobyspring.splearn.domain.MemberStatus;
+import tobyspring.splearn.domain.member.MemberStatus;
 
 class MemberRegisterManualTest {
