+++ src/test/java/tobyspring/splearn/application/required/MemberRepositoryTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.dao.DataIntegrityViolationException;
-import tobyspring.splearn.domain.Member;
+import tobyspring.splearn.domain.member.Member;
 
 @DataJpaTest
 class MemberRepositoryTest {
