+++ src/test/java/tobyspring/splearn/domain/EmailTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
+import tobyspring.splearn.domain.shared.Email;
 
 class EmailTest {
